LPS-61952 Rename com.liferay.portlet.dynamicdatamapping to com.liferay.dynamic.data.mapping.kernel
