+++ portal-service/src/com/liferay/dynamic/data/mapping/kernel/DDMStructureLinkManagerUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.dynamicdatamapping;
+package com.liferay.dynamic.data.mapping.kernel;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.util.ProxyFactory;
