+++ portal-service/src/com/liferay/dynamic/data/mapping/kernel/DDMTemplate.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.dynamicdatamapping;
+package com.liferay.dynamic.data.mapping.kernel;
 
 import com.liferay.portal.model.StagedGroupedModel;
